# swagger/ChatStack.yaml
openapi: 3.0.1
info:
  title: ChatStack API
  version: "1.0"
  description: Chatting Platform
...
servers:
  - url: /chat-stack/api/v1
security:
  - bearerAuth: []

# ===========================================================================
# COMPONENTS
# ===========================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - fullname
        - display_name
        - country
        - email_verified
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: Must be a valid email address
        fullname:
          type: string
        display_name:
          type: string
        email_verified:
          type: boolean
        avatar_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [online, offline]
        timezone:
          type: string
          description: IANA timezone
          example: Europe/Paris
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    UserListResponse:
      type: object
      required:
        - content
        - total_elements
        - total_pages
        - page
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total_elements:
          type: integer
        total_pages:
          type: integer
        page:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          description: >-
            HTTP error or free text such as Database Not Available,
            Invalid Credential, Mandatory, Invalid Format, Invalid Length,
            ForbiddenValue, Invalid Data, etc.
        code:
          type: string
          description: Server / SQL error ID
        description:
          type: string
          description: Human-readable error description including data and value

    # -------------------------------------------------------------------------
    # Auth Request Bodies
    # -------------------------------------------------------------------------
...
    UpdateUserRequest:
      type: object
      description: Fields to update for the current user. All fields are optional.
      properties:
        fullname:
          type: string
          minLength: 2
          maxLength: 255
          example: John Doe
        display_name:
          type: string
          minLength: 3
          maxLength: 50
          pattern: "^[a-zA-Z0-9._-]+$"
          example: john_doe
        country:
          type: string
          minLength: 2
          maxLength: 255
          example: Lebanon
        timezone:
          type: string
          description: IANA timezone
          example: America/New_York
        avatar_url:
          type: string
          nullable: true
          example: "https://cdn.example.com/avatars/john.png"
    AdminUpdateUserRequest:
      type: object
      description: Fields an admin can patch on any user account.
      properties:
        fullname:
          type: string
          minLength: 2
          maxLength: 255
        display_name:
          type: string
          minLength: 3
          maxLength: 50
        country:
          type: string
          minLength: 2
          maxLength: 255
          example: Lebanon
        email_verified:
          type: boolean
        status:
          type: string
          enum: [online, offline]
        role:
          type: string
          enum: [USER, ADMIN]
        timezone:
          type: string
paths:
...
  /users/{id}:
    get:
      operationId: getUserById
      tags:
        - Users Management
      summary: Get any user by ID (public profile)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/avatar/{avatar}:
    get:
      operationId: getUserAvatar
      tags:
        - Users Management
      summary: Get a user's avatar image
      description: >
        Returns the user's avatar image. If no custom avatar is set,
        a default avatar is returned.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: avatar
          in: path
          required: true
          description: Avatar file name (e.g. default.png)
          schema:
            type: string
            example: default.png
      responses:
        '200':
          description: Avatar image binary
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: User or avatar not found
        '410':
          description: Avatar is no longer available

  # ---------------------------------------------------------------------------
  # /users/me — Current authenticated user
  # ---------------------------------------------------------------------------
  /users/me:
    get:
      operationId: getCurrentUser
      tags:
        - Users Management
      summary: Get the current authenticated user
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized — missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User account no longer exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      operationId: patchCurrentUser
      tags:
        - Users Management
      summary: Partially update the current user (only send fields to change)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User patched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ValidationUtils error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: display_name already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: deleteCurrentUser
      tags:
        - Users Management
      summary: Delete the current user account (soft or hard depending on policy)
      responses:
        '204':
          description: Account deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # Admin endpoints
  # ---------------------------------------------------------------------------
  /admin/users:
    get:
      operationId: getAllUsers
      tags:
        - Users Management
      summary: List all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden — admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{id}:
    patch:
      operationId: adminPatchUser
      tags:
        - Users Management
      summary: Partially update any user's data (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: ValidationUtils error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===========================================================================
  # AUTHENTICATION FLOW
  # ===========================================================================
...